requirements.txt
pillow
django==1.7.8
requests
Celery
# It'd be nice to consolidate these, but till then this is how each is used
# used to convert the country name to the currency name
py-moneyed
# used to convert currencies between each other using exchange rates to and from BTC
# This should be updated to have seperate BUY and SELL rates as well I think
money
# we need to check every now and then to see when django-money supports django 1.8 and stick to that version
django-money
# Django countries does the CountryField support
django-countries
djangorestframework
# django-phonenumber-field
django-axes
django-debug-toolbar
# Commented out for now to simply getting setup on windows. Later we will add it when we add Gem support
# round
django-twilio
rapidsms
# this branch included a fix that we needed
git+https://github.com/caktus/rapidsms-twilio.git@develop
# Handles the problem of getting the default currency for a given nation
pycountry
django-subdomains
# Not sure we need either right now
# django-organizations
# django-mptt
django-extensions
#django-organizations
# git+https://alexhayes@bitbucket.org/alexhayes/django-google-places.git
# python-google-places
# https://github.com/slimkrazy/python-google-places/zipball/master
django-celery
PyMySQL
git+https://github.com/Audakel/rapidsms-telerivet.git